feat: add a sidebarIsOpen and onSidebarIsOpen props
The sidebarIsOpen prop contains the current state value as a boolean of whether the sidebar is open or not.

The sidebar Open prop also determines what classes style names will be implemented on the sidebar component.

And the handleSidebarIsOpen will be used for handle toggle open and close the sidebar component.